updates to Auth
